use of new version of counter

+++ PACom/src/main/java/org/proteored/pacom/analysis/exporters/tasks/JTableLoader.java
@@ -279,7 +279,7 @@ public class JTableLoader extends SwingWorker<Void, Void> implements Exporter<JT
 								final List<String> lineStringList = ExporterUtil
 										.getInstance(idSets, includePeptides, retrieveFromUniprot)
 										.getProteinInfoList(proteinGroupOccurrence, columnsStringList,
-												counter.getCount() + 1, idSet);
+												Long.valueOf(counter.getCount()).intValue() + 1, idSet);
 								counter.increment();
 								// System.out.println(peptideString);
 								addNewRow(lineStringList);
